requirements.txt
websockets>=12.0
pygame-ce>=2.5.3
pygame-gui>=0.6.9
